refactor: schedule domain dto record로 변경 및 데이터 정렬 (#112)
