+++ src/main/java/kr/ai/nemo/group/controller/GroupController.java
@@ -12,8 +12,8 @@ import kr.ai.nemo.group.dto.request.GroupSearchRequest;
 import kr.ai.nemo.group.service.GroupCommandService;
 import kr.ai.nemo.group.service.GroupGenerateService;
 import kr.ai.nemo.group.service.GroupQueryService;
-import kr.ai.nemo.schedule.dto.PageRequestDto;
-import kr.ai.nemo.schedule.dto.ScheduleListResponse;
+import kr.ai.nemo.schedule.dto.request.PageRequestDto;
+import kr.ai.nemo.schedule.dto.response.ScheduleListResponse;
 import kr.ai.nemo.schedule.service.ScheduleQueryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
