+++ src/main/java/kr/ai/nemo/schedule/service/ScheduleCommandService.java
@@ -5,8 +5,8 @@ import kr.ai.nemo.group.validator.GroupValidator;
 import kr.ai.nemo.groupparticipants.validator.GroupParticipantValidator;
 import kr.ai.nemo.schedule.domain.Schedule;
 import kr.ai.nemo.schedule.domain.enums.ScheduleStatus;
-import kr.ai.nemo.schedule.dto.ScheduleCreateRequest;
-import kr.ai.nemo.schedule.dto.ScheduleCreateResponse;
+import kr.ai.nemo.schedule.dto.request.ScheduleCreateRequest;
+import kr.ai.nemo.schedule.dto.response.ScheduleCreateResponse;
 import kr.ai.nemo.schedule.validator.ScheduleValidator;
 import kr.ai.nemo.scheduleparticipants.service.ScheduleParticipantsService;
 import kr.ai.nemo.schedule.repository.ScheduleRepository;
